disable caching logic temporarily
Student sessions are ending before other student sessions. So the recalculate condition is causing those later sessions to not be included in the stored results. The logic that Stella explained to me means that there's never truly an end to a class. A user could technically come back a week later and redo an activity, so we'd always have to check if the session was included in the room calculation.