perf: way too many calls to the user content score db
The number of calls was proportional to the number of xapi events.
For an example query in one of the QA accounts:
Before -> 126 calls
After -> 8 calls.

Also added the other two primary keys as query params in order to take advantage of indexing.

Eventually performing the content_id migration will help get rid of
a lot of the ugly code dealing with checking whether it's dealing with an h5p id or not.